Fit of Normal and Exponential-Normal Mixtures to Binned Data

Fit mixture distributions to binned data with a maximum likelihood method, inspired by Venables and Ripley (2002)

fit_unimix(
  breaks,
  counts,
  parms,
  type = c("en", "enn", "ennn", "ennnn", "ennnnn", "n", "nn", "nnn", "nnnn", "nnnnn"),
  sd_min = 0,
  ...
)

fit_n(breaks, counts, parms, sd_min = 0, ...)

fit_en(breaks, counts, parms, sd_min = 0, ...)

fit_nn(breaks, counts, parms, sd_min = 0, ...)

fit_enn(breaks, counts, parms, sd_min = 0, ...)

fit_nnn(breaks, counts, parms, sd_min = 0, ...)

fit_nnnn(breaks, counts, parms, sd_min = 0, ...)

fit_ennn(breaks, counts, parms, sd_min = 0, ...)

fit_ennnn(breaks, counts, parms, sd_min = 0, ...)

fit_nnnnn(breaks, counts, parms, sd_min = 0, ...)

fit_ennnnn(breaks, counts, parms, sd_min = 0, ...)

Arguments

breaks
upper class limits of the data

counts
frequency of observations

parms
list of initial parameters for the mixture

type
of the mixture distribution, e.g. 'enn' for exponential-normal-normal

sd_min
lower boundary value for standard deviation and rate parameter

...
additional arguments passed to mle2

Value

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0

Bolker, Ben and R Development Core Team (2017) bbmle: Tools for General Maximum Likelihood Estimation. R package version 1.0.20. https://CRAN.R-project.org/package=bbmle